+++ src/test/java/io/github/astrapi69/crypto/simple/CharacterSetDecryptorTest.java
@@ -1,0 +1,59 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.crypto.simple;
+
+import static org.testng.Assert.assertEquals;
+
+import java.util.List;
+
+import org.testng.annotations.Test;
+
+/**
+ * The unit test class for the class {@link CharacterSetDecryptor}
+ */
+public class CharacterSetDecryptorTest
+{
+
+	/**
+	 * Test method for the {@link CharacterSetDecryptor#decrypt(List)} method
+	 *
+	 * @throws Exception
+	 *             is thrown if a security error occurs
+	 */
+	@Test
+	public void testDecrypt() throws Exception
+	{
+		String actual;
+		String expected;
+		String text = "Lorem ipsum dolor sit amet, sea consul verterem perfecto id. Alii prompta electram te nec, at minimum copiosae quo. Eos iudico nominati oportere ei, usu at dicta legendos. In nostrum insolens disputando pro, iusto equidem ius id.";
+		List<Character> uniqueCharacters = CharacterSetCrypt.newCharacterList(text);
+		List<Integer> indexesList = CharacterSetCrypt.toIndexList(text, uniqueCharacters);
+		CharacterSetDecryptor decryptor = new CharacterSetDecryptor(uniqueCharacters);
+		decryptor.decrypt(indexesList);
+		actual = decryptor.decrypt(indexesList);
+		expected = text;
+		assertEquals(actual, expected);
+	}
+}